requirements.txt
requests
tweepy
pillow
json
random